fixed full-width urls in metadata. Added some template variable prefixes in components. Took Audioplayer out of AudioDetail page. Wip
