+++ src/components/results/detail/MediaHeader.vue
@@ -4,14 +4,14 @@
     <v-row test-class="subheader">
       <v-col>
         <div class="text-caption mb-n8 text-truncate">
-          <span v-if="file['last-seen']">
-            Last seen <timeago :datetime="file['last-seen']" />
+          <span v-if="$props.file['last-seen']">
+            Last seen <timeago :datetime="$props.file['last-seen']" />
           </span>
-          <span v-if="file.size"> | Size {{ file.size | prettyBytes }}</span>
-          <span> | Mimetype {{ file['mimetype'] }}</span>
+          <span v-if="$props.file.size"> | Size {{ $props.file.size | prettyBytes }}</span>
+          <span> | Mimetype {{ $props.file['mimetype'] }}</span>
           <DownloadButton
-            :hash="file.hash"
-            :title="file.title"
+            :hash="$props.file.hash"
+            :title="$props.file.title"
           />
         </div>
       </v-col>
@@ -22,7 +22,7 @@
       <v-col>
         <div
           class="text-h6 font-weight-regular"
-          v-html="`${file.title || '(no title)'}`"
+          v-html="`${$props.file.title || '(no title)'}`"
         />
       </v-col>
     </v-row>
