+++ src/components/results/detail/MetaDataPanel.vue
@@ -1,5 +1,4 @@
 <template>
-  <!-- Meta data -->
   <v-row>
     <v-col>
       <div class="text-body-1">
@@ -13,8 +12,8 @@
               </template>
             </v-expansion-panel-header>
             <v-expansion-panel-content class="ml-n2 mr-0">
-              <v-simple-table>
-                <template v-slot:default>
+              <template v-slot:default>
+                <v-simple-table>
                   <thead>
                     <tr>
                       <th class="text-left">
@@ -27,15 +26,23 @@
                   </thead>
                   <tbody>
                     <tr
-                      v-for="item in metaData"
+                      v-for="item in metadata"
                       :key="item.key"
                     >
                       <td>{{ item.key }}</td>
                       <td v-html="item.value" />
                     </tr>
+                  </tbody>
+                </v-simple-table>
+                <v-simple-table>
+                  <thead>
                     <tr>
-                      <td>Referenced in:</td>
+                      <th>
+                        Referenced in:
+                      </th>
                     </tr>
+                  </thead>
+                  <tbody>
                     <tr
                       v-for="(item, index) in references"
                       :key="index"
@@ -49,8 +56,8 @@
                       </td>
                     </tr>
                   </tbody>
-                </template>
-              </v-simple-table>
+                </v-simple-table>
+              </template>
             </v-expansion-panel-content>
           </v-expansion-panel>
         </v-expansion-panels>
@@ -68,9 +75,11 @@ export default {
     },
   },
   computed: {
-    metaData() {
-      const metaData = [];
-      // TODO: filters on size, firstseen, lastseen, etc. Make pretty names (e.g. remove hyphens).
+    metadata() {
+      const metadata = [];
+      // TODO: filters on size, firstseen, lastseen, etc.
+      // TODO: Make pretty names (e.g. remove hyphens).
+      // TODO: See if all fields are there.
       [
         'author',
         'creation_date',
@@ -80,9 +89,9 @@ export default {
         'size',
         'type',
       ].forEach((key) => {
-        if (this.file[key]) metaData.push({ key, value: this.file[key] });
+        if (this.file[key]) metadata.push({ key, value: this.file[key] });
       });
-      return metaData;
+      return metadata;
     },
     references() {
       const references = [];
